refactor: 패키지 변경
controller -> config
