+++ src/main/java/roomescape/config/WebMvcConfiguration.java
@@ -5,7 +5,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.controller.LoginMemberArgumentResolver;
 import roomescape.service.MemberService;
 
 @Configuration
