+++ src/main/java/roomescape/controller/ReservationController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.config.LoginMemberArgument;
 import roomescape.domain.Member;
 import roomescape.dto.request.MemberReservationCreationRequest;
 import roomescape.dto.response.ReservationResponse;
